Rename package to eqfiddle; fix sbt build
The code base still derives from the HDES event structure tool and the CoupledSim tool.
It could not be build anymore because the CodeMirror ScalaJS fascade repo went down in 2021.
With the change, the names are more relevant and the build should now work on computers
that do not accidentally have the right dependencies in their cache.